refactor: split datalog engine and add CLI
