+++ .gitignore
@@ -198,7 +198,7 @@ $RECYCLE.BIN/
 *.msp
 
 # Generated Nearley grammar
-src/datalog/datalog-grammar.*
+**/datalog-grammar.*
 
 # Windows shortcuts
 *.lnk
