+++ packages/datalog-cli/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@md-logic/datalog-cli",
+  "bin": {
+    "datalog": "./src/index.ts"
+  },
+  "scripts": {
+    "start": "node --import tsx ./src/index.ts",
+    "test": "echo 'no tests'"
+  },
+  "dependencies": {
+    "@md-logic/datalog": "workspace:*"
+  },
+  "devDependencies": {
+    "@types/node": "^22.15.2",
+    "tsx": "^4.19.3",
+    "typescript": "^5.8.3"
+  }
+}
