+++ packages/datalog/src/datalog.spec.ts
