+++ packages/datalog/src/index.ts
