+++ packages/datalog/src/lexer.ts
